refactor: separate api domain and actors
